+++ src/BigCommerce/Editor/Editor_Dialog_Template.php
@@ -30,7 +30,7 @@ class Editor_Dialog_Template {
 	 * @filter bigcommerce/admin/js_config
 	 */
 	public function js_config( $config, Products_Controller $products_controller, Shortcode_Controller $shortcode_controller ) {
-		$config['editor_dialog'] = [
+		$config[ 'editor_dialog' ] = [
 			'product_api_url'   => $products_controller->get_base_url(),
 			'shortcode_api_url' => $shortcode_controller->get_base_url(),
 		];
@@ -58,40 +58,84 @@ class Editor_Dialog_Template {
 
 	public function render_dialog() {
 		return $this->render_template( 'admin-dialog.php', [
-			'query_builder_sidebar' => $this->query_builder(),
-			'query_settings_sidebar'      => $this->query_settings(),
+			'query_builder_sidebar'  => $this->query_builder(),
+			'query_settings_sidebar' => $this->query_settings(),
 		] );
 	}
 
 	private function query_builder() {
 
-		$featured   = get_term_by( 'slug', Flag::FEATURED, Flag::NAME );
-		$sale       = get_term_by( 'slug', Flag::SALE, Flag::NAME );
-		$brands     = get_terms( [
+		$featured = get_term_by( 'slug', Flag::FEATURED, Flag::NAME );
+		$sale     = get_term_by( 'slug', Flag::SALE, Flag::NAME );
+		$brands   = get_terms( [
 			'taxonomy' => Brand::NAME,
 			'orderby'  => 'title',
 			'order'    => 'ASC',
+		] );
 
-		] );
+		$brands_by_parent = [];
+
+		foreach ( $brands as $brand ) {
+			$brands_by_parent[ $brand->parent ][] = $brand;
+		}
+
+		//  Nest sub-category under the correct parent
+		$brand_choices = $this->get_choices( 0, $brands_by_parent, 0 );
+
 		$categories = get_terms( [
-			'taxonomy' => Product_Category::NAME,
-			'orderby'  => 'title',
-			'order'    => 'ASC',
-			'parent'   => 0,
+			'taxonomy'   => Product_Category::NAME,
+			'orderby'    => 'title',
+			'order'      => 'ASC',
+			'hide_empty' => false,
 		] );
+
+		$categories_by_parent = [];
+
+		foreach ( $categories as $category ) {
+			$categories_by_parent[ $category->parent ][] = $category;
+		}
+
+		//  Nest sub-category under the correct parent
+		$category_choices = $this->get_choices( 0, $categories_by_parent, 0 );
 
 		return $this->render_template( 'query-builder.php', [
 			'featured'   => $featured,
 			'sale'       => $sale,
-			'brands'     => $brands,
-			'categories' => $categories,
+			'brands'     => $brand_choices,
+			'categories' => $category_choices,
 		] );
+	}
+
+	/**
+	 * Reorder the terms so sub-categories get nested bellow its parent
+	 *
+	 * @param $parent_id
+	 * @param $terms_by_parent
+	 * @param $depth
+	 *
+	 * @return array
+	 */
+	private function get_choices( $parent_id, $terms_by_parent, $depth ) {
+		if ( empty( $terms_by_parent[ $parent_id ] ) ) {
+			return [];
+		}
+
+		$choices = [];
+		foreach ( $terms_by_parent[ $parent_id ] as $term ) {
+			$choices[ $term->term_id ][ 'name' ]  = $term->name;
+			$choices[ $term->term_id ][ 'slug' ]  = $term->slug;
+			$choices[ $term->term_id ][ 'id' ]    = $term->term_id;
+			$choices[ $term->term_id ][ 'depth' ] = $depth;
+			$choices = $choices + $this->get_choices( $term->term_id, $terms_by_parent, $depth + 1 );
+		}
+
+		return $choices;
 	}
 
 	private function query_settings() {
 		return $this->render_template( 'query-settings.php', [
 			'posts_per_page' => absint( get_option( Product_Archive::PER_PAGE, Product_Archive::PER_PAGE_DEFAULT ) ),
-		]);
+		] );
 	}
 
 	private function render_template( $filename, $vars ) {