+++ src/BigCommerce/Import/Runner/Lock.php
@@ -4,18 +4,22 @@
 namespace BigCommerce\Import\Runner;
 
 
+use BigCommerce\Import\No_Cache_Options;
+
 class Lock {
+	use No_Cache_Options;
+
 	const OPTION = 'bigcommerce_import.lock';
 
 	public function get_lock() {
-		return get_option( self::OPTION, 0 );
+		return (float) $this->get_option( self::OPTION, 0 );
 	}
 
 	public function release_lock() {
-		update_option( self::OPTION, 0, false );
+		$this->update_option( self::OPTION, 0, false );
 	}
 
 	public function set_lock() {
-		update_option( self::OPTION, microtime( true ), false );
+		$this->update_option( self::OPTION, microtime( true ), false );
 	}
 }